// src/components/BasicModals/BasicModals.less.d.ts
// This file is automatically generated.
// Please do not change this file!
interface CssExports {
  'basic_cancel': string;
  'basic_container': string;
  'basic_content': string;
  'basic_ok': string;
  'close': string;
  'confirm_cancel': string;
  'confirm_ok': string;
  'footer': string;
  'header': string;
  'icon_warning': string;
  'space': string;
  'warn_title': string;
}
export const cssExports: CssExports;
export default cssExports;